GURU2 20190708 첫번 째

+++ app/src/main/java/com/example/semiprojectsample/bean/MemoBean.java
@@ -2,7 +2,7 @@ package com.example.semiprojectsample.bean;
 
 public class MemoBean {
 
-    public int memoId; // 메모 고유 넘버
+    public long memoId; // 메모 고유 넘버
     public String memo;
     public String memoPicPath;
     public String memoDate;
